requirements.txt
requests>=2.27.1
pillow>=9.0.0
pyinstaller>=5.6.2 